[Modify] createUser action
