+++ handler.js
@@ -5,6 +5,7 @@ const db_data = require("./config/db");
 
 const getFunc = require("./lib/get");
 const getAllFunc = require("./lib/getAll");
+const createFunc = require("./lib/create");
 
 module.exports.getUser = async (event, context, callback) => {
   context.callbackWaitsForEmptyEventLoop = false;
@@ -20,36 +21,8 @@ module.exports.getAllUsers = async (event, context, callback) => {
 
 module.exports.createUser = async (event, context, callback) => {
   context.callbackWaitsForEmptyEventLoop = false;
-  let connection;
-  try {
-    connection = await mysql.createConnection(db_data);
-
-    const object = JSON.parse(event.body);
-    const keys = Object.keys(object);
-    const values = keys.map(key => object[key]);
-    const cols = keys.join(", ");
-    const placeholders = values.map(_v => "?").join(", ");
-    const sql = `INSERT into users(${cols}) values(${placeholders})`;
-
-    await connection.query(sql, values);
-    const [result] = await connection.query(
-      "SELECT * FROM users WHERE id = LAST_INSERT_ID()"
-    );
-    connection.end();
-    callback(null, {
-      statusCode: 200,
-      headers: { "Content-type": "application/json" },
-      body: JSON.stringify(result[0])
-    });
-  } catch (err) {
-    connection.end();
-    callback(null, {
-      statusCode: err.statusCode || 500,
-      headers: { "Content-type": "application/json" },
-      body: "Could not create Users: " + err
-    });
-    connection.end();
-  }
+  const response = await createFunc(event, "users");
+  callback(null, response);
 };
 
 module.exports.updateUser = async (event, context, callback) => {
